pyproject.toml
[project]
name = "stock-sentiment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "box>=0.1.5",
    "bs4>=0.0.2",
    "ensure>=1.0.4",
    "feedparser>=6.0.12",
    "ipykernel>=7.1.0",
    "ipywidgets>=8.1.8",
    "langchain>=1.0.4",
    "langchain-community>=0.4.1",
    "langchain-core>=1.0.3",
    "langchain-ollama>=1.0.0",
    "lxml-html-clean>=0.4.3",
    "newsdataapi>=0.1.28",
    "newspaper3k>=0.2.8",
    "ollama>=0.6.0",
    "openllm>=0.6.30",
    "pandas>=2.3.3",
    "parquet>=1.3.1",
    "plotly>=6.4.0",
    "python-box>=7.3.2",
    "requests>=2.32.5",
    "scikit-learn>=1.7.2",
    "streamlit>=1.51.0",
    "torch>=2.9.0",
    "tqdm>=4.67.1",
    "transformers>=4.57.1",
]
